requirements.txt
pandas
numpy
opencv-python
facenet-pytorch
matplotlib
scipy
scikit-image
seaborn
scikit-learn==0.23.2
jupyterlab
pyvi
beautifulsoup4==4.9.3
PyYAML
flask
Flask-Cors